Add an ARM MD flag to give Thumb2 APCS-like frames, saving FP and LR.
Anything that tries to read where the saved PC and SP would be will be
confused, but there are applications (e.g., Linux's perf(1)) that use
only the saved LR and FP.  This is much more efficient than trying to
construct (and tear down) a full APCS frame using only valid Thumb2
instructions would be.

Note that the HARD_FRAME_POINTER_REG is still whatever it would have
been normally, and stack/arg accesses won't be eliminated to FP (which
is good, because that would unnecessarily turn them all into 32-bit
instructions).